backend/ai_agents/generative/requirements-gpu.txt
torch
torchvision
opencv-python
pillow
numpy
diffusers
transformers
accelerate
safetensors
simple-lama
paddleocr
paddlepaddle
easyocr
